tids are long

+++ dyninstAPI/tests/src/test14.mutatee.c
@@ -81,14 +81,14 @@ void level1()
 
    if (i == NTHRD)
    {
-      fprintf(stderr, "[%s:%d] - Error, couldn't find thread id %u\n",
-              __FILE__, __LINE__, (unsigned) me);
+      fprintf(stderr, "[%s:%d] - Error, couldn't find thread id %lu\n",
+              __FILE__, __LINE__, me);
       exit(1);
    }
    if (thrds[i].is_in_instr)
    {
-      fprintf(stderr, "[%s:%d] - Error, thread %u reentered instrumentation\n",
-              __FILE__, __LINE__, (unsigned) me);
+      fprintf(stderr, "[%s:%d] - Error, thread %lu reentered instrumentation\n",
+              __FILE__, __LINE__, me);
       exit(1);
    }
